+++ backend/src/main/kotlin/com/gijun/backend/domain/model/order/OrderItemOption.kt
@@ -4,10 +4,10 @@ import org.springframework.data.annotation.Id
 
 data class OrderItemOption(
     @Id
-    val id: Integer,
-    val orderId: Integer,
-    val orderItemId: Integer,
-    val optionItemId: Integer,
+    val id: Long,
+    val orderId: Long,
+    val orderItemId: Long,
+    val optionItemId: Long,
     val name: String,
     val price: Double,
     val createdAt: String,
