+++ backend/src/main/kotlin/com/gijun/backend/domain/model/store/MenuCategory.kt
@@ -5,9 +5,9 @@ import org.springframework.data.annotation.Id
 
 data class MenuCategory(
     @Id
-    val id: Integer,
-    val storeId: Integer,
-    val displayOrder: Integer,
+    val id: Long,
+    val storeId: Long,
+    val displayOrder: Long,
     val name: String,
     val description: String,
     val activatedStatus: ActivatedStatus,
